backend_clip/requirements.txt
fastapi
uvicorn[standard]
supabase
python-dotenv
Pillow
requests
numpy
scikit-learn
httpx
google-cloud-aiplatform
